Avoid binary font asset by using system OG fonts
