+++ src/components/ThemeProvider.astro
@@ -1,50 +1,27 @@
 ---
-import { themeCssVars } from "@/config/theme";
+const lightVars = `--color-global-bg: oklch(96% 0.018 220);\n--color-global-text: oklch(34% 0.03 232);\n--color-link: oklch(52% 0.085 220);\n--color-accent: oklch(60% 0.09 218);\n--color-accent-2: oklch(38% 0.045 232);\n--color-quote: oklch(48% 0.055 220);\n--color-bg-overlay: rgba(236, 245, 248, 0.78);\n--color-surface: rgba(255, 255, 255, 0.7);`;
+const darkVars = `--color-global-bg: oklch(24% 0.02 232);\n--color-global-text: oklch(90% 0.018 225);\n--color-link: oklch(76% 0.09 220);\n--color-accent: oklch(73% 0.08 215);\n--color-accent-2: oklch(90% 0.02 230);\n--color-quote: oklch(78% 0.05 218);\n--color-bg-overlay: rgba(10, 20, 28, 0.65);\n--color-surface: rgba(17, 28, 37, 0.62);`;
 ---
 
-{/* Inlined to avoid FOUC. This is a parser blocking script. */}
-<script define:vars={{ darkVars: themeCssVars.dark, lightVars: themeCssVars.light }} is:inline>
+<script define:vars={{ darkVars, lightVars }} is:inline>
 	const lightModePref = window.matchMedia("(prefers-color-scheme: light)");
-
-	function getStoredTheme() {
-		return typeof localStorage !== "undefined" ? localStorage.getItem("theme") : null;
-	}
-
-	function getUserPref() {
-		const storedTheme = getStoredTheme();
-		return storedTheme || (lightModePref.matches ? "light" : "dark");
-	}
-
-	function applyThemeVars(root, theme) {
+	const getStoredTheme = () => (typeof localStorage !== "undefined" ? localStorage.getItem("theme") : null);
+	const getUserPref = () => getStoredTheme() || (lightModePref.matches ? "light" : "dark");
+	const applyThemeVars = (root, theme) => {
 		root.style.cssText += theme === "dark" ? darkVars : lightVars;
-	}
-
-	function setTheme(newTheme, persist = true) {
-		if (newTheme !== "light" && newTheme !== "dark") {
-			return console.warn(
-				`Invalid theme value '${newTheme}' received. Expected 'light' or 'dark'.`,
-			);
-		}
-
+	};
+	const setTheme = (newTheme, persist = true) => {
+		if (newTheme !== "light" && newTheme !== "dark") return;
 		const root = document.documentElement;
 		if (newTheme === root.getAttribute("data-theme")) return;
-
 		root.setAttribute("data-theme", newTheme);
 		applyThemeVars(root, newTheme);
-
-		if (persist && typeof localStorage !== "undefined") {
-			localStorage.setItem("theme", newTheme);
-		}
-	}
+		if (persist && typeof localStorage !== "undefined") localStorage.setItem("theme", newTheme);
+	};
 
 	setTheme(getUserPref(), false);
-
 	document.addEventListener("astro:after-swap", () => setTheme(getUserPref(), false));
-
-	document.addEventListener("theme-change", (e) => {
-		setTheme(e.detail.theme, true);
-	});
-
+	document.addEventListener("theme-change", (e) => setTheme(e.detail.theme, true));
 	lightModePref.addEventListener("change", (e) => {
 		if (!getStoredTheme()) setTheme(e.matches ? "light" : "dark", false);
 	});
