+++ src/components/blog/TwikooComments.astro
@@ -1,5 +1,5 @@
 ---
-import { siteSettings } from "@/config/site-settings";
+import { siteSettings } from "@/settings/site";
 ---
 
 <div class="mt-12">
