+++ src/components/layout/Header.astro
@@ -2,159 +2,50 @@
 import Search from "@/components/Search.astro";
 import ThemeToggle from "@/components/ThemeToggle.astro";
 import { menuLinks } from "@/site.config";
-import { siteConfig } from "../../site.config";
+import { siteConfig } from "@/site.config";
 ---
 
-<header class="group relative mb-10 flex items-center justify-between sm:mb-20" id="main-header">
-	<div class="flex items-center">
-		<a
-			aria-current={Astro.url.pathname === "/" ? "page" : false}
-			class="inline-flex items-center"
-			href="/"
-		>
+<header class="mb-10 sm:mb-14" id="main-header">
+	<div class="mb-4 text-center">
+		<a aria-current={Astro.url.pathname === "/" ? "page" : false} class="inline-flex items-center" href="/">
 			<span class="text-xl font-bold sm:text-2xl">{siteConfig.title}</span>
 		</a>
-		<nav
-			aria-label="Main menu"
-			class="bg-global-bg/95 absolute -inset-x-4 top-12 hidden flex-col px-2 py-4 shadow-md backdrop-blur-sm group-[.menu-open]:z-50 group-[.menu-open]:flex sm:static sm:z-auto sm:-ms-4 sm:mt-1 sm:flex sm:flex-row sm:bg-transparent sm:p-0 sm:shadow-none sm:backdrop-blur-none dark:bg-zinc-900 sm:dark:bg-transparent"
-			id="navigation-menu"
-		>
-			{
-				menuLinks.map((link, index) =>
-					link.children?.length ? (
-						<div class="relative px-2 py-4 font-semibold sm:px-4 sm:py-0">
-							<button
-								aria-expanded="false"
-								class="text-accent cursor-pointer"
-								data-submenu-toggle={index}
-								type="button"
-							>
-								{link.title} ▾
-							</button>
-							<div
-								class="submenu bg-global-bg mt-2 hidden flex-col gap-2 sm:absolute sm:start-0 sm:top-full sm:min-w-52 sm:rounded-md sm:p-2 sm:shadow-sm dark:bg-zinc-900"
-								data-submenu={index}
-							>
-								{link.children.map((item) => (
-									<a
-										class="text-accent rounded px-2 py-1 hover:bg-sky-100/60 dark:hover:bg-sky-900/40"
-										href={item.path}
-									>
-										{item.title}
-									</a>
-								))}
-							</div>
-						</div>
-					) : (
-						<a
-							aria-current={Astro.url.pathname === link.path ? "page" : false}
-							class="text-accent px-2 py-4 font-semibold sm:px-4 sm:py-0 sm:underline-offset-2 sm:hover:underline"
-							href={link.path}
-							target={link.external ? "_blank" : undefined}
-							rel={link.external ? "noreferrer" : undefined}
-						>
-							{link.title}
-						</a>
-					),
-				)
-			}
-		</nav>
 	</div>
-	<div class="flex items-center">
+	<div class="flex items-center justify-center gap-2">
+		<nav aria-label="Main menu" class="hidden flex-wrap items-center justify-center gap-2 sm:flex" id="navigation-menu">
+			{menuLinks.map((link, index) =>
+				link.children?.length ? (
+					<div class="relative">
+						<button aria-expanded="false" class="glass-panel card-chip cursor-pointer" data-submenu-toggle={index} type="button">
+							{link.title} ▾
+						</button>
+						<div class="submenu glass-panel absolute left-1/2 top-full z-10 mt-2 hidden min-w-44 -translate-x-1/2 rounded-md p-2">
+							{link.children.map((item) => (
+								<a class="card-chip block" href={item.path}>{item.title}</a>
+							))}
+						</div>
+					</div>
+				) : (
+					<a aria-current={Astro.url.pathname === link.path ? "page" : false} class="glass-panel card-chip" href={link.path} target={link.external ? "_blank" : undefined} rel={link.external ? "noreferrer" : undefined}>{link.title}</a>
+				),
+			)}
+		</nav>
 		<Search />
 		<ThemeToggle />
-		<mobile-button>
-			<button
-				aria-expanded="false"
-				aria-haspopup="menu"
-				class="group relative ms-3 h-7 w-7 sm:invisible sm:hidden"
-				id="toggle-navigation-menu"
-				type="button"
-			>
-				<span class="sr-only">Open main menu</span>
-				<svg
-					aria-hidden="true"
-					class="absolute start-1/2 top-1/2 h-full w-full -translate-x-1/2 -translate-y-1/2 transition-all group-aria-expanded:scale-0 group-aria-expanded:opacity-0"
-					fill="none"
-					focusable="false"
-					id="line-svg"
-					stroke="currentColor"
-					stroke-width="1.5"
-					viewBox="0 0 24 24"
-					xmlns="http://www.w3.org/2000/svg"
-				><path d="M3.75 9h16.5m-16.5 6.75h16.5" stroke-linecap="round" stroke-linejoin="round"
-				></path></svg
-				>
-				<svg
-					aria-hidden="true"
-					class="text-accent absolute start-1/2 top-1/2 h-full w-full -translate-x-1/2 -translate-y-1/2 scale-0 opacity-0 transition-all group-aria-expanded:scale-100 group-aria-expanded:opacity-100"
-					fill="none"
-					focusable="false"
-					id="cross-svg"
-					stroke="currentColor"
-					stroke-width="1.5"
-					viewBox="0 0 24 24"
-					xmlns="http://www.w3.org/2000/svg"
-				><path d="M6 18L18 6M6 6l12 12" stroke-linecap="round" stroke-linejoin="round"></path></svg
-				>
-			</button>
-		</mobile-button>
 	</div>
 </header>
 
 <script>
-	import { toggleClass } from "@/utils/domElement";
-
-	class MobileNavBtn extends HTMLElement {
-		#menuOpen: boolean = false;
-		connectedCallback() {
-			if (this.dataset.bound === "1") return;
-			this.dataset.bound = "1";
-			const headerEl = document.getElementById("main-header")!;
-			const mobileButtonEl = this.querySelector<HTMLButtonElement>("button");
-			mobileButtonEl?.addEventListener("click", () => {
-				if (headerEl) toggleClass(headerEl, "menu-open");
-				this.#menuOpen = !this.#menuOpen;
-				mobileButtonEl.setAttribute("aria-expanded", this.#menuOpen.toString());
-			});
-		}
-	}
-	if (!customElements.get("mobile-button")) customElements.define("mobile-button", MobileNavBtn);
-
 	const initSubmenus = () => {
-		const closeAll = () => {
-			document
-				.querySelectorAll<HTMLElement>("[data-submenu]")
-				.forEach((m) => m.classList.add("hidden"));
-			document
-				.querySelectorAll<HTMLButtonElement>("[data-submenu-toggle]")
-				.forEach((b) => b.setAttribute("aria-expanded", "false"));
-		};
-
-		document.querySelectorAll<HTMLButtonElement>("[data-submenu-toggle]").forEach((btn) => {
+		document.querySelectorAll("[data-submenu-toggle]").forEach((btn) => {
 			if (btn.dataset.bound === "1") return;
 			btn.dataset.bound = "1";
-			btn.addEventListener("click", (e) => {
-				e.preventDefault();
-				e.stopPropagation();
-				const idx = btn.dataset.submenuToggle;
-				const target = document.querySelector<HTMLElement>(`[data-submenu='${idx}']`);
-				const open = target && !target.classList.contains("hidden");
-				closeAll();
-				if (target && !open) {
-					target.classList.remove("hidden");
-					btn.setAttribute("aria-expanded", "true");
-				}
+			btn.addEventListener("click", () => {
+				const submenu = btn.parentElement?.querySelector(".submenu");
+				submenu?.classList.toggle("hidden");
 			});
 		});
-
-		document.addEventListener("click", (e) => {
-			const target = e.target as Node;
-			const inside = (target as Element)?.closest?.("[data-submenu], [data-submenu-toggle]");
-			if (!inside) closeAll();
-		});
 	};
-
 	document.addEventListener("astro:page-load", initSubmenus);
 	initSubmenus();
 </script>
