--- src/config/site-settings.ts
@@ -1,15 +1,0 @@
-export const siteSettings = {
-	font: {
-		bodyClass: "font-mono",
-	},
-	comments: {
-		twikoo: {
-			envId: "https://your-twikoo-env-id.example.com",
-			scriptSrc: "https://cdn.staticfile.net/twikoo/1.6.41/twikoo.all.min.js",
-		},
-	},
-	statusCafe: {
-		username: "isntbunny",
-		atomUrl: "https://status.cafe/users/isntbunny.atom",
-	},
-} as const;
