+++ src/content/note/index.md
@@ -1,0 +1,7 @@
+---
+title: Hello Note
+description: note
+publishDate: "2025-01-01T00:00:00+08:00"
+---
+
+First note.
