+++ src/pages/guestbook.astro
@@ -1,6 +1,6 @@
 ---
 import { getEntry, render } from "astro:content";
-import { siteSettings } from "@/config/site-settings";
+import { siteSettings } from "@/settings/site";
 import PageLayout from "@/layouts/Base.astro";
 
 const meta = {
